Make separate functions for downloading and open in new tab and implement these globally
