+++ src/components/multipleImageModal.js
@@ -6,7 +6,6 @@ import { connect } from 'react-redux'
 import { IMAGE_EXTENSIONS, ITEM_TYPE } from '../constants'
 import css from './css/carousel.css'
 import CustomCarousel from './customCarousel'
-import { handleDownload } from '../helpers/helperfunctions'
 
 class MultipleImageModal extends Component {
   constructor(props) {
@@ -48,6 +47,16 @@ class MultipleImageModal extends Component {
       })
     }
   }
+  handleDownload = () => {
+    const { activeIndex, images } = this.state
+    const item = images[activeIndex]
+    let link = document.createElement('a')
+    link.download = item.fileName
+    link.href = item.upload
+    document.body.appendChild(link)
+    link.click()
+    document.body.removeChild(link)
+  }
 
   render() {
     const { show, onHide, activeItems } = this.props
@@ -83,7 +92,7 @@ class MultipleImageModal extends Component {
             size='large'
             styleName='css.download-icon'
             onClick={() => {
-              handleDownload()
+              this.handleDownload()
             }}
           />
           <CarouselProvider
