+++ src/helpers/helperfunctions.js
@@ -1,6 +1,6 @@
 const Big = require('big.js')
 
-import { ONE_GB, ONE_KB, ONE_MB ,ITEM_TYPE} from '../constants'
+import { ONE_GB, ONE_KB, ONE_MB, ITEM_TYPE } from '../constants'
 import store from '../store'
 
 const removeTrailingZeros = str => {
@@ -61,12 +61,25 @@ export const createContextFromEvent = e => {
   }
 }
 
-export const handleDownload = () => {
+export const openInNewTab = () => {
   const activeItems = store.getState().items.activeItems
   for (const item of activeItems) {
     if (item.type === ITEM_TYPE.file) {
       let link = document.createElement('a')
       link.setAttribute('target', '_blank')
+      link.href = item.obj.upload
+      document.body.appendChild(link)
+      link.click()
+      document.body.removeChild(link)
+    }
+  }
+}
+export const handleDownload = () => {
+  const activeItems = store.getState().items.activeItems
+  for (const item of activeItems) {
+    if (item.type === ITEM_TYPE.file) {
+      let link = document.createElement('a')
+      link.download = item.fileName
       link.href = item.obj.upload
       document.body.appendChild(link)
       link.click()
